Move panels/feeds.channels.ts to common/feed-channels.ts
